join main program and drawing tool into one
