+++ .idea/workspace.xml
@@ -5,6 +5,10 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="5b96f894-a983-4f00-9d42-ceb9bbc5ed88" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/src/main/java/gui/Cords.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/gui/Point.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/java/gui/Subsoil.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/resources/test_changed_map.txt" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.gradle/7.4/executionHistory/executionHistory.bin" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/7.4/executionHistory/executionHistory.bin" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.gradle/7.4/executionHistory/executionHistory.lock" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/7.4/executionHistory/executionHistory.lock" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.gradle/7.4/fileHashes/fileHashes.bin" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/7.4/fileHashes/fileHashes.bin" afterDir="false" />
@@ -13,18 +17,10 @@
       <change beforePath="$PROJECT_DIR$/.gradle/buildOutputCleanup/buildOutputCleanup.lock" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/buildOutputCleanup/buildOutputCleanup.lock" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.gradle/buildOutputCleanup/outputFiles.bin" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/buildOutputCleanup/outputFiles.bin" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.gradle/file-system.probe" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/file-system.probe" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/gradle.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/gradle.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/build/resources/main/table pow.txt" beforeDir="false" afterPath="$PROJECT_DIR$/build/resources/main/table pow.txt" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/build/resources/main/table.txt" beforeDir="false" afterPath="$PROJECT_DIR$/build/resources/main/table.txt" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/build/tmp/compileJava/previous-compilation-data.bin" beforeDir="false" afterPath="$PROJECT_DIR$/build/tmp/compileJava/previous-compilation-data.bin" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Program.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/Program.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/drawingtool/DrawingTool.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/drawingtool/DrawingTool.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/drawingtool/auxiliary/Board.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/drawingtool/auxiliary/Board.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/drawingtool/auxiliary/Subsoil.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/drawingtool/auxiliary/Subsoil.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/java/gui/Board.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/gui/Board.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/resources/table pow.txt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/table pow.txt" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/resources/table.txt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/table.txt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/gui/GUI.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/gui/GUI.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -74,19 +70,23 @@
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent">
-    <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
-    <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
-    <property name="SHARE_PROJECT_CONFIGURATION_FILES" value="true" />
-    <property name="last_opened_file_path" value="$PROJECT_DIR$" />
-    <property name="project.structure.last.edited" value="Modules" />
-    <property name="project.structure.proportion" value="0.15" />
-    <property name="project.structure.side.proportion" value="0.2" />
-  </component>
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "WebServerToolWindowFactoryState": "false",
+    "node.js.detected.package.eslint": "true",
+    "node.js.detected.package.tslint": "true",
+    "node.js.selected.package.eslint": "(autodetect)",
+    "node.js.selected.package.tslint": "(autodetect)"
+  }
+}]]></component>
   <component name="RecentsManager">
     <key name="MoveClassesOrPackagesDialog.RECENTS_KEY">
       <recent name="main" />
       <recent name="main.drawingtool" />
+    </key>
+    <key name="CopyClassDialog.RECENTS_KEY">
+      <recent name="main.java.gui" />
+      <recent name="main.java.gui.actors" />
     </key>
   </component>
   <component name="RunManager" selected="Application.Program">
@@ -150,6 +150,7 @@
       <option name="presentableId" value="Default" />
       <updated>1652733937451</updated>
       <workItem from="1652733938812" duration="922000" />
+      <workItem from="1653231281064" duration="2898000" />
     </task>
     <servers />
   </component>
